// CodePulse.UI/src/app/admin/blog-post/models/blog-post.ts
import { Category } from '../../category/models/category';

export interface BlogPost {
  id?: string;
  title: string;
  shortDescription: string;
  content: string;
  featuredImageUrl: string;
  urlHandle: string;
  author: string;
  publishedDate: Date;
  isVisible: boolean;
  categories: Category[];
}